move the propagation targets to utils
